replaced the entra id support with a client app
