--- entra-id-users-support-quickstart/.env.example
@@ -1,9 +1,0 @@
-# Azure Communication Services resource endpoint URI
-COMMUNICATION_SERVICES_RESOURCE_ENDPOINT=https://<your-resource>.communication.azure.com
-
-# Microsoft Entra client ID and tenant ID values
-ENTRA_CLIENT_ID=<your-client-id>
-ENTRA_TENANT_ID=<your-tenant-id>
-
-# Server port
-SERVER_PORT="3000"
