+++ entra-id-users-support-quickstart/package.json
@@ -3,16 +3,20 @@
   "version": "1.0.0",
   "main": "index.js",
   "scripts": {
+    "dev": "vite",
+    "build": "vite build",
+    "preview": "vite preview",
     "test": "echo \"Error: no test specified\" && exit 1"
   },
   "keywords": [],
   "author": "",
-  "license": "ISC",
+  "license": "MIT",
   "description": "",
   "dependencies": {
-    "@azure/communication-common": "^2.3.2-beta.1",
-    "@azure/identity": "^4.10.0",
-    "dotenv": "^16.5.0",
-    "express": "^5.1.0"
+    "@azure/communication-common": "^2.4.0",
+    "@azure/identity": "^4.10.1",
+    "react": "^19.1.0",
+    "react-dom": "^19.1.0",
+    "vite": "^6.3.5"
   }
 }
